Added replanner and major code refactoring
