+++ src/Planner.cpp
@@ -15,7 +15,7 @@ Planner::Planner(void)
 	// create a goal state
 	ob::ScopedState<ob::RealVectorStateSpace> goal(space);
 
-	// set the bounds for the R^3 part of SE(3)
+	// set the bounds for the R^3
 	ob::RealVectorBounds bounds(3);
 
 	bounds.setLow(0,-10);
@@ -23,7 +23,7 @@ Planner::Planner(void)
 	bounds.setLow(1,-10);
 	bounds.setHigh(1,10);
 	bounds.setLow(2,0);
-	bounds.setHigh(2,10);
+	bounds.setHigh(2,5);
 
 	space->as<ob::RealVectorStateSpace>()->setBounds(bounds);
 
@@ -67,22 +67,25 @@ Planner::~Planner()
 {
 }
 
-void Planner::init_start(void)
-{
-	if(!set_autostart)
-		DBG("Initialized");
-	set_autostart = true;
-}
-void Planner::setStart(double x, double y, double z)
+bool Planner::setStart(double x, double y, double z)
 {
 	ob::ScopedState<ob::RealVectorStateSpace> start(space);
 	// start->setXYZ(x,y,z);
 	start->values[0] = x;
 	start->values[1] = y;
 	start->values[2] = z;
-	// start->as<ob::SO3StateSpace::StateType>(1)->setIdentity();
-	pdef->clearStartStates();
-	pdef->addStartState(start);
+	ob::State *state =  space->allocState();
+	state->as<ob::RealVectorStateSpace::StateType>()->values = start->values;
+	if(isStateValid(state)) // Check if the start state is valid
+	{	
+		// start->as<ob::SO3StateSpace::StateType>(1)->setIdentity();
+		pdef->clearStartStates();
+		pdef->addStartState(start);
+		DBG("Start point set to: " << x << " " << y << " " << z);
+		return true;
+	}
+	else
+		return false;
 }
 void Planner::setGoal(double x, double y, double z)
 {
@@ -100,40 +103,51 @@ void Planner::setGoal(double x, double y, double z)
 		pdef->clearGoal();
 		pdef->setGoalState(goal);
 		DBG("Goal point set to: " << x << " " << y << " " << z);
-		if(set_autostart)
-			plan();
-		
+
 	}
 }
-void Planner::updateMap(octomap::OcTree* tree_oct)
+void Planner::updateMap(octomap::OcTree tree_oct)
 {
 	// convert octree to collision object
-	fcl::OcTree<double>* tree = new fcl::OcTree<double>(std::shared_ptr<const octomap::OcTree>(tree_oct));
+	fcl::OcTree<double>* tree = new fcl::OcTree<double>(std::make_shared<const octomap::OcTree>(tree_oct));
 	std::shared_ptr<fcl::CollisionGeometry<double>> tree_obj = std::shared_ptr<fcl::CollisionGeometry<double>>(tree);
 	treeObj = std::make_shared<fcl::CollisionObject<double>>((tree_obj));
 }
-void Planner::replan(void)
-{
-	if(path_smooth != NULL && set_autostart)
+bool Planner::replan(void)
+{	
+	if(path_smooth != NULL)
 	{
 		DBG("Total Points:" << path_smooth->getStateCount ());
-		if(path_smooth->getStateCount () <= 2)
-			plan();
+		double distance;
+		if(pdef->hasApproximateSolution())
+		{
+			DBG("Goal state not satisfied and distance to goal is: " << pdef->getSolutionDifference());
+			replan_flag = true;
+		}
 		else
 		{
 			for (std::size_t idx = 0; idx < path_smooth->getStateCount (); idx++)
 			{
 				if(!replan_flag)
+				{
+					// const ob::RealVectorStateSpace::StateType *pos = path_smooth->getState(idx)->as<ob::RealVectorStateSpace::StateType>();
 					replan_flag = !isStateValid(path_smooth->getState(idx));
+				}
 				else
 					break;
-
 			}
-			if(replan_flag)
-				plan();
-			else
-				DBG("Replanning not required");
 		}
+	}
+	if(replan_flag)
+	{
+		DBG("Replanning");
+		plan();
+		return true;
+	}
+	else
+	{
+		DBG("Replanning not required");
+		return false;
 	}
 }
 
@@ -150,13 +164,13 @@ void Planner::plan(void)
 	plan->setup();
 
     // print the settings for this space
-	si->printSettings(std::cout);
+	// si->printSettings(std::std::cout);
 
     // print the problem settings
-	pdef->print(std::cout);
+	// pdef->print(std::cout);
 
-    // attempt to solve the problem within one second of planning time
-	ob::PlannerStatus solved = plan->solve(2);
+    // attempt to solve the problem within three seconds of planning time
+	ob::PlannerStatus solved = plan->solve(3);
 
 	if (solved)
 	{
@@ -168,13 +182,12 @@ void Planner::plan(void)
 		pth->printAsMatrix(std::cout);
         // print the path to screen
         // path->print(std::cout);
-		// path_smooth = pth;
 		
         //Path smoothing using bspline
 
 		og::PathSimplifier* pathBSpline = new og::PathSimplifier(si);
 		path_smooth = new og::PathGeometric(dynamic_cast<const og::PathGeometric&>(*pdef->getSolutionPath()));
-		pathBSpline->smoothBSpline(*path_smooth,3);
+		pathBSpline->smoothBSpline(*path_smooth);
 
 		// DBG("Smoothed Path");
 		// path_smooth.print(std::cout);
@@ -195,7 +208,7 @@ bool Planner::isStateValid(const ob::State *state)
 
     // check validity of state defined by pos
 	fcl::Vector3<double> translation(pos->values[0],pos->values[1],pos->values[2]);
-
+	// INFO("State: " << translation);
 	aircraftObject->setTranslation(translation);
 	fcl::CollisionRequest<double> requestType(1,false,1,false);
 	fcl::CollisionResult<double> collisionResult;
@@ -208,12 +221,6 @@ bool Planner::isStateValid(const ob::State *state)
 // for optimal motion planning. This method returns an objective which
 // attempts to minimize the length in configuration space of computed
 // paths.
-ob::OptimizationObjectivePtr Planner::getThresholdPathLengthObj(const ob::SpaceInformationPtr& si)
-{
-	ob::OptimizationObjectivePtr obj(new ob::PathLengthOptimizationObjective(si));
-	// obj->setCostThreshold(ob::Cost(1.51));
-	return obj;
-}
 
 ob::OptimizationObjectivePtr Planner::getPathLengthObjWithCostToGo(const ob::SpaceInformationPtr& si)
 {
